edit scenarios

+++ db/migrate/20190415230150_create_scenario_weights.rb
@@ -3,7 +3,7 @@ class CreateScenarioWeights < ActiveRecord::Migration[5.2]
     create_table :scenario_weights do |t|
       t.references :company, null: false
       t.references :scenario, null: false
-      t.integer :weight, default: 1
+      t.float :weight, default: 1
 
       t.timestamps
     end
